rewrite to to reduce bucket to just key and value
looking more closely at Google's sparsehash, I realized the reason that
one needs to set an empty and deleted key is because there is no
additional variables used in each bucket for bucket status. after
thinking some more, I agree with this approach because a cache miss is
much more detrimental in terms of speed, so introducing the fact that
you have to manually set sentinel keys is worth it.

unfortunately, since I wanted to support non-POD data types, this is a
lot more complicated.

refactor macro names to start with KOKOPUFFS prefix.

add stubs for min_load_factor, to be worked on later.